+++ packages/syncfusion_flutter_pdfviewer/CHANGELOG.md
@@ -1,4 +1,9 @@
-## [unreleased]
+## [19.1.57-beta] - 04/20/2021
+
+* Added headers parameter in `SfPdfViewer.network`.
+* Now, PDF document will be displayed properly inside the `SizedBox` or `Container` Widgets in Web platform.
+
+## [19.1.55-beta] - 04/06/2021
 
 **Features**